+++ libs/rudder-integration-firebase-react-native/ios/RudderIntegrationFirebaseReactNative.h
@@ -1,5 +1,0 @@
-#import <React/RCTBridgeModule.h>
-
-@interface RudderIntegrationFirebaseReactNative : NSObject <RCTBridgeModule>
-
-@end