+++ libs/rudder-integration-firebase-react-native/CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [2.0.0](https://github.com/rudderlabs/rudder-sdk-react-native/compare/rudder-integration-firebase-react-native@1.2.0...rudder-integration-firebase-react-native@2.0.0) (2025-06-05)
+
+
+### ⚠ BREAKING CHANGES
+
+* **firebase:** add support for new react native architecture (#450)
+
+### Features
+
+* **firebase:** add support for new react native architecture ([#450](https://github.com/rudderlabs/rudder-sdk-react-native/issues/450)) ([4fe8551](https://github.com/rudderlabs/rudder-sdk-react-native/commit/4fe8551a952f910c825c27be2cc5027d6803d75c))
+
 ## [1.2.0](https://github.com/rudderlabs/rudder-sdk-react-native/compare/rudder-integration-firebase-react-native@1.1.0...rudder-integration-firebase-react-native@1.2.0) (2025-04-29)